Made trivia use Option and multiline functions
Use Option for Scores and QuestionList, which may not be initialized when the TriviaManager is created.

They are set to Some when Start is called

Also made chained methods easier to read by putting them on multiple lines.
This is important because, with using option, there is no a lot of "as_ref" and "expect" calls