Added Channel to OptionSet, rearranged code
Added a channel to the OptionSet struct. Now, when a new game is started, the optionset will contain the channel where the game is being played, and only outputs messages to that channel.

Because of this change, all of the initialization code for a "start" command is located in the TriviaManager. Before, the channel was set in the on_message handler in another section of code.

Also made sure all relevant TriviaManager members were destructed (set to None, in this case) when a game ends.