Added multichoice listing for questions
Moved away from short answer format to a multichoice format, where the answers are displayed in an A-D fashion, and answers are given in A-D form as well.

The current plan is to make it one guess per user per question, to avoid speed guessing all possible letters. This is not implimented yet